A first implementation of the generalised normal distribution was included.
This follows broadly the Wikipedia page for notation - this differs slightly
from the implementation in scipy.stats, but the quantitative results are
identical as far as I can see.

This only implements the symmetric distribution - in accorandance with the
naming of normal distributions in pymc I have opted against including
Symmetric in the name. It would also be moderately easy to implement the
asymmetric distribution based on what is included here.